feat: add tailwindcss
This commit adds Tailwind CSS to the project.

Tailwind CSS is a utility-first CSS framework that allows developers to quickly and easily create custom designs. It is a great choice for projects that need a lot of customization, such as this one.

The following changes were made to add Tailwind CSS to the project:

- A `tailwind.config.js` file was added to the project root. This file contains the configuration for Tailwind CSS.
- The `postcss.config.js` file was updated to include the Tailwind CSS plugin.
- The `src/assets/global.css` file was added to the project. This file contains the global styles for the project.
- The `src/lib/components/hero.svelte` file was updated to use Tailwind CSS.
- The `src/routes/+layout.svelte` file was updated to include the Tailwind CSS stylesheet.
- The `svelte.config.js` file was updated to include the Vite plugin for Svelte. This plugin allows Tailwind CSS to be used with Svelte.