feat: add JogjaJS footer
This commit adds a new footer to the JogjaJS website. The footer includes the JogjaJS logo, links to the website's social media pages, and a copyright notice.

The footer is designed to be responsive and work well on all devices. It is also styled to match the rest of the website's design.

This change improves the overall look and feel of the website and makes it more consistent with other JogjaJS branding.